@@ -20,11 +20,11 @@
     "@google-cloud/bigquery": "^5.11.0",
     "@kinecosystem/kin-sdk-v2": "^0.5.0",
     "@nestjs/apollo": "^10.0.6",
-    "@nestjs/common": "^8.4.0",
+    "@nestjs/common": "9.0.7",
     "@nestjs/config": "^1.2.0",
-    "@nestjs/core": "^8.4.0",
+    "@nestjs/core": "9.0.7",
     "@nestjs/graphql": "^10.0.6",
-    "@nestjs/platform-express": "^8.4.0",
+    "@nestjs/platform-express": "9.0.7",
     "@nestjs/schedule": "^1.0.2",
     "@nestjs/serve-static": "^2.2.2",
     "@solana/web3.js": "^1.42.0",
@@ -45,31 +45,28 @@
     "tslib": "^2.3.1"
   },
   "devDependencies": {
-    "@nestjs/schematics": "^8.0.8",
-    "@nestjs/testing": "^8.4.0",
-    "@nrwl/cli": "13.8.5",
-    "@nrwl/cypress": "13.8.5",
-    "@nrwl/eslint-plugin-nx": "13.8.5",
-    "@nrwl/jest": "13.8.5",
-    "@nrwl/js": "13.8.5",
-    "@nrwl/linter": "13.8.5",
-    "@nrwl/nest": "13.8.5",
-    "@nrwl/node": "13.8.5",
-    "@nrwl/nx-cloud": "13.1.6",
-    "@nrwl/tao": "13.8.5",
-    "@nrwl/web": "13.8.5",
-    "@nrwl/workspace": "13.8.5",
-    "@tailwindcss/forms": "^0.5.0",
-    "@tailwindcss/typography": "^0.5.2",
+    "@nestjs/schematics": "9.0.1",
+    "@nestjs/testing": "9.0.7",
+    "@nrwl/cli": "14.5.2",
+    "@nrwl/cypress": "14.5.2",
+    "@nrwl/eslint-plugin-nx": "14.5.2",
+    "@nrwl/jest": "14.5.2",
+    "@nrwl/js": "14.5.2",
+    "@nrwl/linter": "14.5.2",
+    "@nrwl/nest": "14.5.2",
+    "@nrwl/node": "14.5.2",
+    "@nrwl/nx-cloud": "14.3.0",
+    "@nrwl/web": "14.5.2",
+    "@nrwl/workspace": "14.5.2",
     "@types/fs-extra": "^9.0.13",
     "@types/jest": "27.4.1",
-    "@types/node": "17.0.21",
-    "@typescript-eslint/eslint-plugin": "5.14.0",
-    "@typescript-eslint/parser": "5.14.0",
+    "@types/node": "18.0.0",
+    "@typescript-eslint/eslint-plugin": "5.32.0",
+    "@typescript-eslint/parser": "5.32.0",
     "autoprefixer": "^10.4.2",
     "babel-jest": "27.5.1",
     "cypress": "^9.5.1",
-    "eslint": "8.10.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.5.0",
     "eslint-plugin-cypress": "^2.12.1",
     "eslint-plugin-import": "2.25.4",
@@ -77,10 +74,11 @@
     "husky": "^7.0.4",
     "jest": "27.5.1",
     "lint-staged": "^12.3.5",
-    "prettier": "2.5.1",
-    "ts-jest": "27.1.3",
+    "nx": "14.5.2",
+    "prettier": "2.7.1",
+    "ts-jest": "27.1.4",
     "ts-morph": "^14.0.0",
-    "typescript": "~4.5.5"
+    "typescript": "4.7.4"
   },
   "lint-staged": {
     "*.{js,jsx,json,css,scss,md,ts,tsx,html,graphql,yaml}": [
