@@ -1,6 +1,7 @@
-module.exports = {
-  displayName: 'api',
-  preset: '../../jest.preset.js',
+/* eslint-disable */
+export default {
+  displayName: 'api-network-feature',
+  preset: '../../../../jest.preset.js',
   globals: {
     'ts-jest': {
       tsconfig: '<rootDir>/tsconfig.spec.json',
@@ -11,5 +12,5 @@ module.exports = {
     '^.+\\.[tj]s$': 'ts-jest',
   },
   moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../coverage/apps/api',
+  coverageDirectory: '../../../../coverage/libs/api/network/feature',
 }
